Fix a slow redirect issue
Redirecting to the overview page from a very large form resulted in the browser freezing for several seconds. The underlying issue is observer system in the forking-store which gets used by the form fields addon. The latest patch fixes that issue upstream, so we update to that version to also fix the issue here.